1. Ctr-middle / Alt-middle opens Object editor in a new window 2. nsis/nsis-runall.sh is configurable via command-line switches and (Linux only) uploads private installers 3. replace c++11 "override" with WOO_CXX11_OVERRIDE macro, expanding to nothing for g++<=4.6 (so that the compilation does not fail)

+++ nsis/nsis-runall.sh
@@ -12,35 +12,60 @@ else
 	BINDIR=/c/src/woo/dist/wwoo-win64
 fi
 
+DO_LIBS=false
+DO_WOO=false
+DO_EXTRA=false
+DO_UPLOAD=false
+
+while getopts ":lweu" opt; do
+	case $opt in
+		l) DO_LIBS=true ;;
+		w) DO_WOO=true ;;
+		e) DO_EXTRA=true ;;
+		u) DO_UPLOAD=true ;;
+   esac
+done
+
+
 REVNO=`bzr revno ..`
 DESTDIR=$BINDIR/..
 
 cp *.nsh $BINDIR/
 cp *.rtf $BINDIR/
 rm -f $BINDIR/*-installer.exe
-# copy eggs to $BINDIR; they are not installed with other installers, so it is safe to have them there
-rm -f $BINDIR/wooExtra*.egg
-for d in ../wooExtra/*; do
-	echo $d
-	[ -f $d/setup.py ] || continue
-	pushd $d
-		rm -rf dist/*.egg
-		python setup.py bdist_egg
-		cp dist/*.egg $BINDIR
-	popd
-done
+
+if $DO_EXTRA; then
+	# copy eggs to $BINDIR; they are not installed with other installers, so it is safe to have them there
+	rm -f $BINDIR/wooExtra*.egg
+	for d in ../wooExtra/*; do
+		echo $d
+		[ -f $d/setup.py ] || continue
+		pushd $d
+			rm -rf dist/*.egg
+			python setup.py bdist_egg
+			cp dist/*.egg $BINDIR
+		popd
+	done
+fi
+
+echo 'HERE'
 
 if [ -d /boot ]; then
 	# linux
 	pushd $BINDIR
-		#makensis -DVERSION=1.0c nsis-wwoo-libs.nsh 
-		#makensis -DVERSION=0.99-r$REVNO nsis-wwoo-main.nsh 
+		if $DO_LIBS; then makensis -DVERSION=1.0c nsis-wwoo-libs.nsh; fi
+		if $DO_WOO; then makensis -DVERSION=0.99-r$REVNO nsis-wwoo-main.nsh; fi
 		# make installers for extra modules
-		for EGG in wooExtra.*.egg; do
-			COMPONENT=`echo $EGG | cut -f1 -d-`
-			VERSION=`echo $EGG | cut -f2 -d-`
-			makensis -DCOMPONENT=$COMPONENT -DVERSION=$VERSION nsis-wwoo-extra.nsh
-		done
+		if $DO_EXTRA; then
+			for EGG in wooExtra.*.egg; do
+				COMPONENT=`echo $EGG | cut -f1 -d-`
+				VERSION=`echo $EGG | cut -f2 -d-`
+				makensis -DCOMPONENT=$COMPONENT -DVERSION=$VERSION nsis-wwoo-extra.nsh
+				if $DO_UPLOAD; then
+					scp $EGG Woo-$COMPONENT-*-installer.exe bbeta:host/woodem/private/`python -c "import $COMPONENT; print $COMPONENT.KEY"`/inst/
+				fi
+			done
+		fi
 	popd
 else
 	# windows is so singularly ugly
@@ -48,19 +73,24 @@ else
 	pushd $BINDIR
 		# work around msys expanding /D: http://forums.winamp.com/showthread.php?t=253732
 		# omg
-		echo "!define VERSION 1.0c" > defines.nsh
-		"$MAKENSIS" defines.nsh nsis-wwoo-libs.nsh
-		echo "!define VERSION 0.99-r$REVNO" > defines.nsh
-		"$MAKENSIS" defines.nsh nsis-wwoo-main.nsh
-		for EGG in wooExtra.*.egg; do
-			COMPONENT=`echo $EGG | cut -f1 -d-`
-			VERSION=`echo $EGG | cut -f2 -d-`
-			echo "!define COMPONENT $COMPONENT" > defines.nsh
-			echo "!define VERSION $VERSION" >>defines.nsh
-			"$MAKENSIS" defines.nsh nsis-wwoo-extra.nsh
-		done
+		if $DO_LIBS; then
+			echo "!define VERSION 1.0c" > defines.nsh
+			"$MAKENSIS" defines.nsh nsis-wwoo-libs.nsh
+		fi
+		if $DO_WOO; then
+			echo "!define VERSION 0.99-r$REVNO" > defines.nsh
+			"$MAKENSIS" defines.nsh nsis-wwoo-main.nsh
+		fi
+		if $DO_EXTRA; then
+			for EGG in wooExtra.*.egg; do
+				COMPONENT=`echo $EGG | cut -f1 -d-`
+				VERSION=`echo $EGG | cut -f2 -d-`
+				echo "!define COMPONENT $COMPONENT" > defines.nsh
+				echo "!define VERSION $VERSION" >>defines.nsh
+				"$MAKENSIS" defines.nsh nsis-wwoo-extra.nsh
+			done
+		fi
 	popd
 fi
-
 
 mv $BINDIR/*-installer.exe $DESTDIR
